// src/app/e-wallet/models/api/cardOutputModel.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Address } from './address';
import { Card } from './card';
import { EligibleCard } from './eligibleCard';

export interface CardOutputModel { 
    readonly cards?: Array<Card>;
    readonly eligibleCards?: Array<EligibleCard>;
    readonly lastAddress?: Address;
}